Implemented the Eagle class. Eagle moves according to Bresenham's line algorithm. moveEagle and makeMoveToSword/Start are implemented (may need to be modified). Next step should be creating an eagle takeOff function and a moveWithHero function.
Changes needed at Game class:
Make the eagle move along with the hero;
Ask user if he wants to "shoot" eagle;
Make the eagle take off, "telling" her the sword position;
Move the eagle, and comparing with the surroundings to change its state
accordingly (check if its on top of anything);
"Tell" the eagle if the sword is available or not when she drops to get
it (may need an extra variable for the eagle);
"Tell" the eagle when the hero arrives its place (if the eagle is
waiting for him).

User outputs are needed too.